# frontend/src/client/openapi.json
{"openapi": "3.0.2", "info": {"title": "Gummistiefel B", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Hello World", "operationId": "hello_world__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/detail/{id}": {"get": {"summary": "Detail", "operationId": "detail_detail__id__get", "parameters": [{"required": true, "schema": {"title": "Id", "type": "integer"}, "name": "id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/query": {"get": {"summary": "Query", "description": "Query the Database on the Fulltext Search index and return the specified fields\nof all matching documents. Optionally limit the results.\n\n\nExample URL:\n/query?length__lte=1&severity_index__gt=0&fields=area&fields=length&limit=200\n\nWill parse to:\nfilters = [[length, lte, 1], [severity_index, gt, 0]]\nfields  = [area, length]\nlimit   = 200", "operationId": "query_query_get", "parameters": [{"required": false, "schema": {"title": "Filter Params", "type": "string", "default": ""}, "name": "filter_params", "in": "query"}, {"required": false, "schema": {"title": "Fields", "type": "array", "items": {"type": "string"}}, "name": "fields", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer"}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/overview": {"get": {"summary": "Overview", "operationId": "overview_overview_get", "parameters": [{"required": false, "schema": {"title": "Bins", "type": "integer", "default": 20}, "name": "bins", "in": "query"}, {"required": false, "schema": {"title": "Filter Params", "type": "string", "default": ""}, "name": "filter_params", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/spider": {"get": {"summary": "Spider", "operationId": "spider_spider_get", "parameters": [{"required": true, "schema": {"title": "Intervala", "type": "string"}, "name": "intervalA", "in": "query"}, {"required": true, "schema": {"title": "Intervalb", "type": "string"}, "name": "intervalB", "in": "query"}, {"required": false, "schema": {"title": "Filter Params", "type": "string", "default": ""}, "name": "filter_params", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/overview-histogram": {"get": {"summary": "Overview Histogram", "operationId": "overview_histogram_overview_histogram_get", "parameters": [{"required": false, "schema": {"title": "Filter Params", "type": "string", "default": ""}, "name": "filter_params", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}